requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.0.post1
sqlalchemy>=2.0.25
pydantic[email]>=2.5.3
pymysql>=1.1.0
python-dotenv>=1.0.1
